Refactored application.properties api.endpoint property to spring.mvc.servlet.path; moved SecurityConfiguration to separate directory
