+++ src/main/resources/application.properties
@@ -1,5 +1,7 @@
 spring.application.name=libman-api
 
+spring.mvc.servlet.path=/api/v1
+
 # needed for any script-based initialization
 spring.sql.init.mode=always
